+++ tests/test_pytest.py
@@ -101,7 +101,13 @@ def test_wildcard(parser_test_wildcard: igittigitt.IgnoreParser, base_path: path
     assert parser_test_wildcard.match(base_path / "hello.txt")
     assert parser_test_wildcard.match(base_path / "hello.foobar/")
     assert parser_test_wildcard.match(base_path / "dir/hello.txt")
-    assert parser_test_wildcard.match(base_path / "hello.")
+
+    if platform.system() == "Linux":
+        assert parser_test_wildcard.match(base_path / "hello.")
+    else:
+        # in Windows there can be no files ending with a point
+        assert not parser_test_wildcard.match(base_path / "hello.")
+
     assert not parser_test_wildcard.match(base_path / "hello")
     assert not parser_test_wildcard.match(base_path / "helloX")