+++ CHANGES.rst
@@ -11,6 +11,8 @@ v2.1.1b
     - set __all__ accordingly
     - point out limitations in Readme
     - integrate github actions
+    - adjusting tests: patterns ending with a point can not match on windows
+    - adressing invalid escape sequence '\ '
 
 v2.1.0
 ------